requirements.txt
python-telegram-bot>=20.7
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
claude-code-sdk>=0.0.25
openai>=1.0.0
